+++ ModCore/Listeners/Timers.cs
@@ -1,419 +1,44 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
+﻿using ConcurrentCollections;
 using DSharpPlus;
 using DSharpPlus.Entities;
-using DSharpPlus.EventArgs;
-using Humanizer;
+using Microsoft.AspNetCore.Components;
 using Microsoft.Extensions.Logging;
 using ModCore.Database;
 using ModCore.Database.DatabaseEntities;
 using ModCore.Database.JsonEntities;
 using ModCore.Entities;
-using ModCore.Extensions;
 using ModCore.Extensions.Attributes;
 using ModCore.Extensions.Enums;
-using ModCore.Utils;
 using ModCore.Utils.Extensions;
+using Newtonsoft.Json;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace ModCore.Listeners
 {
     public static class Timers
     {
-        //[AsyncListener(EventType.Ready)]
-        public static async Task OnReady(DatabaseContextBuilder database, DiscordClient client, SharedData sharedData)
+        private static DatabaseContextBuilder databaseContextBuilder = null;
+        private static SemaphoreSlim semaphore = new SemaphoreSlim(1, 1);
+        private static (DatabaseTimer timer, CancellationTokenSource cancellation) current = (null, null);
+        private static DiscordClient client = null;
+
+        [AsyncListener(EventType.Ready)]
+        public static async Task OnReadyAsync(DatabaseContextBuilder database, DiscordClient clientInject)
         {
-            using (var db = database.CreateContext())
+            if (databaseContextBuilder == null)
             {
-                if (!db.Timers.Any())
-                    return;
-
-                var guildids = client.Guilds.Select(xg => (long)xg.Key).ToArray();
-                var timers = db.Timers.Where(xt => guildids.Contains(xt.GuildId)).ToArray();
-                if (!timers.Any())
-                    return;
-
-                // lock timers
-                await sharedData.TimerSempahore.WaitAsync();
-                try
-                {
-                    var now = DateTimeOffset.UtcNow;
-                    var pasttimers = timers.Where(xt => xt.DispatchAt <= now.AddSeconds(30)).ToArray();
-                    if (pasttimers.Any())
-                    {
-                        foreach (var timer in pasttimers)
-                        {
-                            // dispatch past timers
-                            _ = DispatchTimer(new TimerData(null, timer, client, database, sharedData, null));
-                        }
-
-                        db.Timers.RemoveRange(pasttimers);
-                        await db.SaveChangesAsync();
-                    }
-                }
-                catch (Exception ex)
-                {
-                    client.Logger.Log(LogLevel.Error, "ModCore",
-                        $"Caught Exception in Timer Ready: {ex.GetType().ToString()}\n{ex.StackTrace}", DateTime.UtcNow);
-                }
-                finally
-                {
-                    // unlock the timers
-                    sharedData.TimerSempahore.Release();
-                }
-
-                await RescheduleTimers(client, database, sharedData);
+                databaseContextBuilder = database;
             }
-        }
-
-        public static async Task TimerCallback(Task task, object state)
-        {
-            var timerdata = state as TimerData;
-            var shard = timerdata.Context;
-            var shared = timerdata.Shared;
-
-            // lock the timers
-            await shared.TimerSempahore.WaitAsync();
-
-            try
+            if (client == null)
             {
-                // remove the timer
-                using (var db = timerdata.Database.CreateContext())
-                {
-                    db.Timers.Remove(timerdata.DbTimer);
-                    db.SaveChanges();
-                }
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Caught Exception in Timer Callback: {ex.GetType().ToString()}\n{ex.StackTrace}");
-            }
-            finally
-            {
-                timerdata.Shared.TimerData = null;
-                // release the lock
-                shared.TimerSempahore.Release();
+                client = clientInject;
             }
 
-            // dispatch the timer
-            _ = Task.Run(LocalDispatch);
-
-            // schedule new one
-            await RescheduleTimers(shard, timerdata.Database, timerdata.Shared);
-
-            Task LocalDispatch()
-                => DispatchTimer(timerdata);
-        }
-
-        private static async Task DispatchTimer(TimerData tdata)
-        {
-            var timer = tdata.DbTimer;
-            var client = tdata.Context;
-            if (timer.ActionType == TimerActionType.Reminder)
-            {
-                DiscordChannel channel = null;
-                DiscordMessage original = null;
-                try
-                {
-                    channel = await client.GetChannelAsync((ulong)timer.ChannelId);
-                }
-                catch
-                {
-                    return;
-                }
-
-                if (channel == null)
-                    return;
-
-                var data = timer.GetData<TimerReminderData>();
-                try
-                {
-                    if (data.MessageId != null)
-                        original = await channel.GetMessageAsync(data.MessageId.Value, true);
-                }
-                catch (Exception) { } // message no longer exists
-
-                var emoji = DiscordEmoji.FromName(client, ":alarm_clock:");
-                var user = (ulong)timer.UserId;
-
-                var message = $"<@{user}>";
-                var unixTimestamp = new DateTimeOffset(timer.DispatchAt);
-
-                var msg = new DiscordMessageBuilder();
-
-                msg.AddEmbed(
-                    new DiscordEmbedBuilder()
-                    .WithTitle("⏰ Reminder")
-                    .WithDescription(data.Snoozed? $"You snoozed a reminder to be re-reminded <t:{unixTimestamp.ToUnixTimeSeconds()}:R>" 
-                        : $"You wanted to be reminded <t:{unixTimestamp.ToUnixTimeSeconds()}:R>")
-                    .AddField("✏️ Reminder Text", data.ReminderText)
-                    );
-
-                msg.WithReply(data.MessageId, true, false);
-
-                // generating a snowflake from original unix timestamp
-                ulong fakeContextId = ((ulong)data.OriginalUnix - 1420070400000ul) << 22;
-
-                var link = string.IsNullOrEmpty(data.SnoozedContext) ? $"https://discord.com/channels/{channel.GuildId}/{channel.Id}/{fakeContextId}" : data.SnoozedContext;
-
-                var snooze = new DiscordButtonComponent(ButtonStyle.Secondary, "snooze", "Snooze", emoji: new DiscordComponentEmoji(DiscordEmoji.FromUnicode("⏰")));
-                var contextLink = new DiscordLinkButtonComponent(link, "Jump to Context");
-
-                msg.AddComponents(snooze, contextLink);
-                msg.WithContent(message);
-                msg.WithAllowedMention(new UserMention(user));
-
-                await msg.SendAsync(channel);
-
-                // ALWAYS filter stuff so i set it to false. No need to @everyone in a reminder.
-            }
-            else if (timer.ActionType == TimerActionType.Unban)
-            {
-                var data = timer.GetData<TimerUnbanData>();
-                if (client.Guilds.Any(x => x.Key == (ulong)timer.GuildId))
-                {
-                    using (var db = tdata.Database.CreateContext())
-                    {
-                        var guild = client.Guilds[(ulong)timer.GuildId];
-                        try
-                        {
-                            await guild.UnbanMemberAsync((ulong)data.UserId);
-                        }
-                        catch
-                        {
-                            // ignored
-                        }
-
-                        var settings = guild.GetGuildSettings(db);
-
-                        var embed = new DiscordEmbedBuilder()
-                            .WithTitle($"Tempban Expired")
-                            .WithDescription($"{data.DisplayName}#{data.Discriminator} has been unbanned. ({data.UserId})")
-                            .WithColor(DiscordColor.Green);
-                        await guild.ModLogAsync(db, embed);
-                    }
-                }
-            }
-            else if (timer.ActionType == TimerActionType.Unmute)
-            {
-                /* --- DEPRECATED --- */
-
-                // DEPRECATED
-                //var data = timer.GetData<TimerUnmuteData>();
-                //if (client.Guilds.Any(x => x.Key == (ulong)timer.GuildId))
-                //{
-                //	using (var db = tdata.Database.CreateContext())
-                //	{
-                //		var guild = client.Guilds[(ulong)timer.GuildId];
-                //		var member = await guild.GetMemberAsync((ulong)data.UserId);
-                //		var role = (DiscordRole)null;
-                //		try
-                //		{
-                //			role = guild.GetRole((ulong)data.MuteRoleId);
-                //		}
-                //		catch (Exception)
-                //		{
-                //			try
-                //			{
-                //				role = guild.GetRole(guild.GetGuildSettings(db).MuteRoleId);
-                //			}
-                //			catch (Exception)
-                //			{
-                //				await client.LogAutoActionAsync(guild, db, $"**[IMPORTANT]**\nFailed to unmute member: {data.DisplayName}#{data.Discriminator} (ID: {data.UserId})\nMute role does not exist!");
-                //				return;
-                //			}
-                //		}
-                //		await member.RevokeRoleAsync(role, "");
-                //		await client.LogAutoActionAsync(guild, db, $"Member unmuted: {data.DisplayName}#{data.Discriminator} (ID: {data.UserId})");
-                //	}
-                //}
-            }
-            else if (timer.ActionType == TimerActionType.Pin)
-            {
-                /* --- DEPRECATED --- */
-
-                //var data = timer.GetData<TimerPinData>();
-                //if (client.Guilds.Any(x => x.Key == (ulong)timer.GuildId))
-                //{
-                //	using (var db = tdata.Database.CreateContext())
-                //	{
-                //		var guild = client.Guilds[(ulong)timer.GuildId];
-                //		var channel = guild.GetChannel((ulong)data.ChannelId);
-                //		var message = await channel.GetMessageAsync((ulong)data.MessageId);
-                //		await message.PinAsync();
-
-                //		var embed = new DiscordEmbedBuilder()
-                //				.WithTitle($"Message pin scheduled")
-                //				.AddField("Channel", $"<#{channel.Id}>")
-                //				.AddField("Message", $"{message.Id}")
-                //				.AddField("Content", $"{message.Content.Truncate(1000)}")
-                //				.WithColor(DiscordColor.Purple);
-                //		await guild.ModLogAsync(db, embed);
-                //	}
-                //}
-            }
-            else if (timer.ActionType == TimerActionType.Unpin)
-            {
-                /* --- DEPRECATED --- */
-
-                //var data = timer.GetData<TimerPinData>();
-                //if (client.Guilds.Any(x => x.Key == (ulong)timer.GuildId))
-                //{
-                //	using (var db = tdata.Database.CreateContext())
-                //	{
-                //		var guild = client.Guilds[(ulong)timer.GuildId];
-                //		var channel = guild.GetChannel((ulong)data.ChannelId);
-                //		var message = await channel.GetMessageAsync((ulong)data.MessageId);
-                //		await message.UnpinAsync();
-                //		var embed = new DiscordEmbedBuilder()
-                //				.WithTitle($"Message unpin scheduled")
-                //				.AddField("Channel", $"<#{channel.Id}>")
-                //				.AddField("Message", $"{message.Id}")
-                //				.AddField("Content", $"{message.Content.Truncate(1000)}")
-                //				.WithColor(DiscordColor.Purple);
-                //		await guild.ModLogAsync(db, embed);
-                //	}
-                //}
-            }
-        }
-
-        public static async Task<TimerData> RescheduleTimers(DiscordClient client, DatabaseContextBuilder database, SharedData shared)
-        {
-            // lock the timers
-            await shared.TimerSempahore.WaitAsync();
-
-            var tdata = shared.TimerData;
-
-            try
-            {
-                // set a new timer
-                DatabaseTimer[] timers = null;
-                bool force = false;
-
-                using (var db = database.CreateContext())
-                {
-                    var ids = client.Guilds.Select(xg => (long)xg.Key).ToArray();
-                    timers = db.Timers.Where(xt => ids.Contains(xt.GuildId))
-                        .OrderBy(xt => xt.DispatchAt)
-                        .ToArray();
-
-                    if (shared.TimerData != null)
-                        force = db.Timers.Count(xt => xt.Id == shared.TimerData.DbTimer.Id) == 0; // .Any() throws
-                }
-
-                var nearest = timers.FirstOrDefault();
-                if (nearest == null)
-                {
-                    // there's no nearest timer
-                    return null;
-                }
-
-                if (tdata != null && tdata.DbTimer.Id == nearest.Id)
-                {
-                    // it's the same timer
-                    return tdata;
-                }
-
-
-                if (CancelIfLaterThan(nearest.DispatchAt, shared, force))
-                {
-                    var cts = new CancellationTokenSource();
-                    var dispatch = nearest.DispatchAt - DateTimeOffset.UtcNow;
-
-                    if (dispatch.TotalMilliseconds > Int32.MaxValue)
-                    {
-                        using (var db = database.CreateContext())
-                        {
-                            db.Timers.Add(new DatabaseTimer()
-                            {
-                                ActionType = TimerActionType.Unknown,
-                                ActionData = "{}",
-                                DispatchAt = DateTime.Now.AddMilliseconds(Int32.MaxValue),
-                                ChannelId = 0,
-                                GuildId = 0,
-                                UserId = 0
-                            });
-                            await db.SaveChangesAsync();
-                        }
-
-                        return await RescheduleTimers(client, database, shared);
-                    }
-
-                    var task = Task.Delay(dispatch, cts.Token);
-                    tdata = new TimerData(task, nearest, client, database, shared, cts);
-                    _ = task.ContinueWith(TimerCallback, tdata, TaskContinuationOptions.OnlyOnRanToCompletion);
-                    shared.TimerData = tdata;
-                }
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Caught Exception in Timer Reschedule: {ex.GetType().ToString()}\n{ex.StackTrace}");
-            }
-            finally
-            {
-                // release the lock
-                shared.TimerSempahore.Release();
-            }
-
-            return tdata;
-        }
-
-        public static async Task<TimerData> UnscheduleTimerAsync(DatabaseTimer timer, DiscordClient shard, DatabaseContextBuilder database, SharedData shared)
-        {
-            // lock the timers
-            await shared.TimerSempahore.WaitAsync();
-
-            try
-            {
-                // remove the requested timer
-                using (var db = database.CreateContext())
-                {
-                    db.Timers.Remove(timer);
-                    await db.SaveChangesAsync();
-                }
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Caught Exception in Timer Unschedule: {ex.GetType().ToString()}\n{ex.StackTrace}");
-            }
-            finally
-            {
-                // release the lock
-                shared.TimerSempahore.Release();
-            }
-
-            // trigger a reschedule
-            return await RescheduleTimers(shard, database, shared);
-        }
-
-        public static async Task<TimerData> UnscheduleTimersAsync(List<DatabaseTimer> timers, DiscordClient shard, DatabaseContextBuilder database, SharedData shared)
-        {
-            // lock the timers
-            await shared.TimerSempahore.WaitAsync();
-
-            try
-            {
-                // remove the requested timers
-                using (var db = database.CreateContext())
-                {
-                    db.Timers.RemoveRange(timers);
-                    await db.SaveChangesAsync();
-                }
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Caught Exception in Timer Unschedule: {ex.GetType().ToString()}\n{ex.StackTrace}");
-            }
-            finally
-            {
-                // release the lock
-                shared.TimerSempahore.Release();
-            }
-
-            // trigger a reschedule
-            return await RescheduleTimers(shard, database, shared);
+            await ScheduleNextAsync();
         }
 
         public static DatabaseTimer FindNearestTimer(TimerActionType actionType, ulong userId, ulong channelId, ulong guildId, DatabaseContextBuilder database)
@@ -428,17 +53,197 @@ namespace ModCore.Listeners
                 return db.Timers.FirstOrDefault(xt => xt.Id == id && xt.ActionType == actionType && xt.UserId == (long)userId);
         }
 
-        private static bool CancelIfLaterThan(DateTimeOffset dto, SharedData shared, bool force)
+        public static async Task UnscheduleTimersAsync(params DatabaseTimer[] timers)
         {
-            // check if a timer is going
-            if (force || shared.TimerData == null || shared.TimerData.DispatchTime >= dto)
+            // lock the timers
+            await semaphore.WaitAsync();
+
+            try
             {
-                var timerdata = shared.TimerData;
-                timerdata?.Cancel?.Cancel();
-                return true;
+                // remove the requested timers
+                using (var db = databaseContextBuilder.CreateContext())
+                {
+                    db.Timers.RemoveRange(timers);
+                    await db.SaveChangesAsync();
+                }
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine($"Caught Exception in Timer Unschedule: {ex.GetType().ToString()}\n{ex.StackTrace}");
+            }
+            finally
+            {
+                // release the lock
+                semaphore.Release();
             }
 
-            return false;
+            await ScheduleNextAsync();
+        }
+
+        public static async Task ScheduleNextAsync()
+        {
+            await semaphore.WaitAsync();
+            await TriggerExpiredTimersAsync();
+
+            using var dbContext = databaseContextBuilder.CreateContext();
+            if (!dbContext.Timers.Any())
+            {
+                return; // We'll reschedule once a new timer arrives through a command.
+            }
+
+            var newTimer = dbContext.Timers.OrderBy(x => x.DispatchAt).FirstOrDefault();
+            if (newTimer != null)
+            {
+                // Cancel previous if necessary
+                if (current.timer == null || current.timer.DispatchAt > newTimer.DispatchAt)
+                {
+                    if (current.cancellation != null)
+                    {
+                        // We cancel our previous dispatch / interim reschedule
+                        current.cancellation.Cancel();
+                    }
+
+                    // Schedule next
+                    current.cancellation = new CancellationTokenSource();
+
+                    var delay = newTimer.DispatchAt.Subtract(DateTime.Now);
+                    // There's a max to a delay. If it's over the max, we reschedule on trigger. Else, just schedule dispatch.
+                    if (delay.TotalMilliseconds > Int32.MaxValue)
+                    {
+                        current.timer = null;
+                        _ = Task.Delay(TimeSpan.FromMilliseconds(Int32.MaxValue - 1000/* just to be safe */), current.cancellation.Token)
+                            .ContinueWith(InterimScheduleNextAsync, current.cancellation, TaskContinuationOptions.OnlyOnRanToCompletion);
+                    }
+                    else
+                    {
+                        current.timer = newTimer;
+                        _ = Task.Delay(delay, current.cancellation.Token)
+                            .ContinueWith(DispatchAsync, current.timer, TaskContinuationOptions.OnlyOnRanToCompletion);
+                    }
+                }
+            }
+            semaphore.Release();
+        }
+
+        private static async Task TriggerExpiredTimersAsync()
+        {
+            using var dbContext = databaseContextBuilder.CreateContext();
+
+            var now = DateTime.UtcNow.AddSeconds(2);
+            var expiredTimers = dbContext.Timers.Where(x => x.DispatchAt < now).ToList();
+
+            foreach (var timer in expiredTimers)
+            {
+                // Dispatch all expired timers without rescheduling.
+                try
+                {
+                    await DispatchAsync(timer, false);
+                }
+                catch (Exception) { }
+            }
+        }
+
+        private static async Task InterimScheduleNextAsync(Task task, object data) => await ScheduleNextAsync();
+
+        private static async Task DispatchAsync(Task task, object timer) => await DispatchAsync((DatabaseTimer)timer, true);
+
+        private static async Task DispatchAsync(DatabaseTimer timer, bool dispatchNext)
+        {
+            switch (timer.ActionType)
+            {
+                default:
+                    client.Logger.LogWarning($"Unknown timer action triggered! {JsonConvert.SerializeObject(timer)}");
+                    break;
+
+                case TimerActionType.Reminder:
+                    await DispatchReminderAsync(timer, timer.GetData<TimerReminderData>());
+                    break;
+
+                case TimerActionType.Unban:
+                    await DispatchUnbanAsync(timer, timer.GetData<TimerUnbanData>());
+                    break;
+            }
+
+            using(var dbContext = databaseContextBuilder.CreateContext())
+            {
+                dbContext.Timers.Remove(timer);
+                await dbContext.SaveChangesAsync();
+            }
+
+            if (dispatchNext)
+            {
+                current.timer = null;
+                current.cancellation = null;
+                await ScheduleNextAsync();
+            }
+        }
+
+        private static async Task DispatchReminderAsync(DatabaseTimer timer, TimerReminderData data)
+        {
+            DiscordChannel channel;
+            try
+            {
+                channel = await client.GetChannelAsync((ulong)timer.ChannelId);
+            }
+            catch (Exception)
+            {
+                return;
+            }
+
+            var mention = $"<@{timer.UserId}>";
+            var unixTimestamp = new DateTimeOffset(timer.DispatchAt);
+            // generating a snowflake from original unix timestamp
+            ulong fakeContextId = ((ulong)data.OriginalUnix - 1420070400000ul) << 22;
+            var link = string.IsNullOrEmpty(data.SnoozedContext) ? $"https://discord.com/channels/{channel.GuildId}/{channel.Id}/{fakeContextId}" : data.SnoozedContext;
+            var snooze = new DiscordButtonComponent(ButtonStyle.Secondary, "snooze", "Snooze", emoji: new DiscordComponentEmoji(DiscordEmoji.FromUnicode("⏰")));
+            var contextLink = new DiscordLinkButtonComponent(link, "Jump to Context");
+
+            var msg = new DiscordMessageBuilder()
+            .AddEmbed
+            (
+                new DiscordEmbedBuilder()
+                .WithTitle("⏰ Reminder")
+                .WithDescription(data.Snoozed ? $"You snoozed a reminder to be re-reminded <t:{unixTimestamp.ToUnixTimeSeconds()}:R>"
+                    : $"You wanted to be reminded <t:{unixTimestamp.ToUnixTimeSeconds()}:R>")
+                .AddField("✏️ Reminder Text", data.ReminderText)
+            )
+            .WithReply(data.MessageId, true, false)
+            .AddComponents(snooze, contextLink)
+            .WithContent(mention)
+            .WithAllowedMention(new UserMention((ulong)timer.UserId));
+
+            await msg.SendAsync(channel);
+        }
+
+        private static async Task DispatchUnbanAsync(DatabaseTimer timer, TimerUnbanData data)
+        {
+            DiscordGuild guild;
+
+            try
+            {
+                guild = await client.GetGuildAsync((ulong)timer.GuildId);
+            }
+            catch(Exception)
+            {
+                return;
+            }
+
+            using (var db = databaseContextBuilder.CreateContext())
+            {
+                try
+                {
+                    await guild.UnbanMemberAsync((ulong)data.UserId);
+                }
+                catch
+                {
+                    // ignored
+                }
+                var embed = new DiscordEmbedBuilder()
+                    .WithTitle($"Tempban Expired")
+                    .WithDescription($"{data.DisplayName}#{data.Discriminator} has been unbanned. ({data.UserId})")
+                    .WithColor(DiscordColor.Green);
+                await guild.ModLogAsync(db, embed);
+            }
         }
     }
 }
